+++ packages/expo-router/CHANGELOG.md
@@ -26,6 +26,8 @@
 ### 💡 Others
 
 - Create href preview component ([#37335](https://github.com/expo/expo/pull/37335) by [@Ubax](https://github.com/Ubax))
+- Add formsheet warning ([#37982](https://github.com/expo/expo/pull/37982) by [@Ubax](https://github.com/Ubax))
+
 
 ## 5.1.3 - 2025-07-03